requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-crispy-forms==1.14.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.2.1
telerivet==1.6.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
